+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 5.2.1 (11/8/2017 - )
 * log: moved message class out of impl, to make log-processor not depending on impl package (prepare for jdk 9)
+* log: log-processor collects its own cpu/heap/kafka stats to index
 
 ### 5.2.0 (10/28/2017 - 11/7/2017)
 * bean: removed "javax.inject:javax.inject:1", replaced with core.framework.inject, removed constructor injection support (to simplify and prepare for JDK 9)